CUSTCOM-211 Upgraded Trilead library and included ECDSA support

+++ nucleus/cluster/admin/src/main/java/com/sun/enterprise/v3/admin/cluster/SecureAdminBootstrapHelper.java
@@ -308,8 +308,8 @@ public abstract class SecureAdminBootstrapHelper {
          * @param milliseconds normal Java time (in milliseconds)
          * @return
          */
-        Integer secondsSince_01_Jan_1970(final long milliseconds) {
-            return (int) (milliseconds) / 1000;
+        long secondsSince_01_Jan_1970(final long milliseconds) {
+            return milliseconds / 1000;
         }
     }
